requirements.txt
# Satori Tiger Legal Document Processing Dependencies

# Core document processing
docling>=1.0.0
python-docx>=0.8.11

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0

# Document processing and OCR
PyPDF2>=3.0.1
pdfplumber>=0.10.0
Pillow>=10.0.0

# Template engine for document generation
Jinja2>=3.1.2

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# System monitoring (for performance tests)
psutil>=5.9.0

# Validation and schemas
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.9.0
flake8>=6.1.0